+++ src/components/question-answer/QuestionAnswer.js
@@ -1,37 +1,16 @@
-/* eslint-disable no-undef */
-import React, { useState, useEffect } from 'react';
-import { connect } from 'react-redux';
-import { useParams } from 'react-router-dom';
-
+import React, { useState } from 'react';
 import SearchBar from './SearchBar';
 import Questions from './Questions';
 
-const getQuestions = (id) => fetch(`http://3.134.102.30/qa/${id}`).then((res) => res.json());
 
-const QuestionAnswer = ({ productStore }) => {
-  const { id } = useParams();
-  const [questions, setQuestions] = useState([]);
-  // looking to see if product has changed or if startup and nothing in questions
-  // eslint-disable-next-line no-undef
-  if (Number(id) !== Number(productStore.id) || questions.length === 0) {
-    getQuestions(id).then((result) => {
-      // update the questions in state
-      setQuestions(result.results);
-    });
-  }
-  return (
-    <div>
-      <h3>
+const QuestionAnswer = () => (
+  <div>
+    <h3>
         QUESTIONS & ANSWER
-      </h3>
-      <SearchBar />
-      <Questions questions={questions} />
-    </div>
-  );
-};
+    </h3>
+    <SearchBar />
+    <Questions />
+  </div>
+);
 
-const mapStateToProps = (state) => ({
-  productStore: state.product,
-});
-
-export default connect(mapStateToProps)(QuestionAnswer);
+export default QuestionAnswer;
